+++ packages/editor/editor-core/src/plugins/layout/styles.ts
@@ -5,6 +5,9 @@ import { colors, gridSize } from '@atlaskit/theme';
 import {
   columnLayoutSharedStyle,
   gridMediumMaxWidth,
+  akEditorDeleteBackground,
+  akEditorDeleteBorder,
+  akEditorDeleteBorderSize,
 } from '@atlaskit/editor-common';
 
 export const layoutStyles = css`
@@ -21,7 +24,7 @@ export const layoutStyles = css`
         padding: ${gridSize() * 1.5}px;
         flex: 1;
         min-width: 0;
-        border: 1px solid transparent;
+        border: 1px solid ${colors.N40};
         border-radius: 5px;
       }
 
@@ -44,7 +47,12 @@ export const layoutStyles = css`
       }
       /* Shows the border when cursor is inside a layout */
       &.selected > * {
-        border-color: ${colors.N40};
+        border-color: ${colors.N50};
+      }
+
+      &.selected.danger > [data-layout-column] {
+        background-color: ${akEditorDeleteBackground};
+        border: ${akEditorDeleteBorderSize}px solid ${akEditorDeleteBorder};
       }
     }
   }