+++ packages/editor/editor-core/src/plugins/media/styles.ts
@@ -1,7 +1,12 @@
 // @ts-ignore: unused variable
 // prettier-ignore
 import { css, Styles, StyledComponentClass } from 'styled-components';
-import { mediaSingleSharedStyle } from '@atlaskit/editor-common';
+import {
+  mediaSingleSharedStyle,
+  akEditorDeleteBorder,
+  akEditorDeleteBorderBoldSize,
+  akEditorDeleteBackground,
+} from '@atlaskit/editor-common';
 import { colors } from '@atlaskit/theme';
 
 export const mediaStyles = css`
@@ -82,6 +87,40 @@ export const mediaStyles = css`
     left: unset !important;
     width: auto !important;
     height: auto !important;
+  }
+
+  /* Danger when top level node for smart cards / inline links */
+  .danger > div > div > .media-card-frame,
+  .danger > span > a {
+    background-color: ${akEditorDeleteBackground};
+    box-shadow: 0px 0px 0px ${akEditorDeleteBorderBoldSize}px
+      ${akEditorDeleteBorder};
+    transition: background-color 0s;
+    transition: box-shadow 0s;
+  }
+
+  /* Danger when nested node or common */
+  .danger {
+    /* Media single */
+    .media-single div div div::after {
+      border: ${akEditorDeleteBorderBoldSize}px solid ${akEditorDeleteBorder};
+    }
+
+    .mediaSingle-selected .mediaSingle-resize-handle-right::after,
+    .mediaSingle-selected .mediaSingle-resize-handle-left::after {
+      background: ${akEditorDeleteBorder};
+    }
+
+    /* Smart cards */
+    div div .media-card-frame,
+    .inlineCardView-content-wrap > span > a {
+      background-color: rgb(255, 189, 173, 0.5); /* R75 with 50% opactiy */
+      transition: background-color 0s;
+    }
+
+    div div .media-card-frame::after {
+      box-shadow: none;
+    }
   }
 `;